fix: 同步订单到管家婆

+++ app/utils/store/goods/dto.go
@@ -49,7 +49,7 @@ type ProductData interface {
 type ProductSku interface {
 	GetSkuId() int64
 	GetSkuName() string
-	GetProductId() int
+	GetProductId() int64
 	GetOuterId() string
 	GetPrice() int
 	GetQuantity() int
